fix(j-s): Adding appeal state in all case list tables  (#11318)
* adding appeal state to appeal

* update naming on get tag

* removing old code

---------
