+++ apps/judicial-system/web/src/routes/CourtOfAppeal/Cases/Cases.tsx
@@ -8,6 +8,7 @@ import {
   PageHeader,
   SectionHeading,
   Table,
+  TagAppealState,
 } from '@island.is/judicial-system-web/src/components'
 import { titles, tables, core } from '@island.is/judicial-system-web/messages'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
@@ -29,9 +30,8 @@ import {
   CaseState,
   isRestrictionCase,
 } from '@island.is/judicial-system/types'
-import { Box, Tag, TagVariant, Text } from '@island.is/island-ui/core'
+import { Box, Text } from '@island.is/island-ui/core'
 import BigTextSmallText from '@island.is/judicial-system-web/src/components/BigTextSmallText/BigTextSmallText'
-import TagAppealRuling from '@island.is/judicial-system-web/src/components/TagAppealRuling/TagAppealRuling'
 import { AppealedCasesQuery } from '@island.is/judicial-system-web/src/utils/mutations'
 
 import { logoContainer } from '../../Shared/Cases/Cases.css'
@@ -176,29 +176,12 @@ const CourtOfAppealCases = () => {
         }
       }) => {
         const thisRow = row.row.original
-        const tagVariant: { color: TagVariant; text: string } =
-          thisRow.appealState === CaseAppealState.APPEALED
-            ? {
-                color: 'purple',
-                text: formatMessage(tables.newTag),
-              }
-            : thisRow.appealState === CaseAppealState.RECEIVED
-            ? { color: 'darkerBlue', text: formatMessage(tables.receivedTag) }
-            : { color: 'darkerBlue', text: formatMessage(tables.completedTag) }
 
         return (
-          <>
-            <Box marginRight={1} marginBottom={1}>
-              <Tag variant={tagVariant.color} outlined disabled>
-                {tagVariant.text}
-              </Tag>
-            </Box>
-            {thisRow.appealState === CaseAppealState.COMPLETED && (
-              <TagAppealRuling
-                appealRulingDecision={thisRow.appealRulingDecision}
-              />
-            )}
-          </>
+          <TagAppealState
+            appealState={thisRow.appealState}
+            appealRulingDecision={thisRow.appealRulingDecision}
+          />
         )
       },
     },
