+++ apps/judicial-system/web/src/routes/Shared/Cases/PastCases.tsx
@@ -25,12 +25,12 @@ import {
 import { useViewport } from '@island.is/judicial-system-web/src/utils/hooks'
 import {
   Table,
+  TagAppealState,
   UserContext,
 } from '@island.is/judicial-system-web/src/components'
 import { core, tables } from '@island.is/judicial-system-web/messages'
 import { CaseType } from '@island.is/judicial-system-web/src/graphql/schema'
 import BigTextSmallText from '@island.is/judicial-system-web/src/components/BigTextSmallText/BigTextSmallText'
-import TagAppealRuling from '@island.is/judicial-system-web/src/components/TagAppealRuling/TagAppealRuling'
 
 import { displayCaseType, mapCaseStateToTagVariant } from './utils'
 import * as styles from './Cases.css'
@@ -211,8 +211,9 @@ const PastCases: React.FC<Props> = (props) => {
                   {tagVariant.text}
                 </Tag>
               </Box>
-              {row.row.original.appealState === CaseAppealState.COMPLETED && (
-                <TagAppealRuling
+              {row.row.original.appealState && (
+                <TagAppealState
+                  appealState={row.row.original.appealState}
                   appealRulingDecision={row.row.original.appealRulingDecision}
                 />
               )}
